refactor(ui): extract GROUP_COLORS and hexToRgba to shared module
建立 modules/ui/groupColors.js 集中管理 Tab Group 顏色常數與 hexToRgba 工具函式，
避免在 tabRenderer.js 和 otherWindowRenderer.js 中重複定義，遵循 DRY 原則。

變更內容：
- 新增 modules/ui/groupColors.js：集中匯出 GROUP_COLORS 和 hexToRgba
- 修改 tabRenderer.js：移除重複定義，改為 import 共享模組
- 修改 otherWindowRenderer.js：同上
- 更新 PRD_spec.md：新增 FR-08、FR-09 說明此需求
- 更新 SA_spec.md：架構圖新增 groupColors.js、新增 2.4 節說明、更新依賴關係
