+++ modules/ui/tabRenderer.js
@@ -2,25 +2,8 @@ import * as api from '../apiManager.js';
 import { ADD_TO_GROUP_ICON_SVG, BOOKMARK_ICON_SVG } from '../icons.js';
 import { tabListContainer } from './elements.js';
 import { showContextMenu } from './contextMenuManager.js';
+import { GROUP_COLORS, hexToRgba } from './groupColors.js';
 
-const GROUP_COLORS = {
-    grey: '#5f6368',
-    blue: '#8ab4f8',
-    red: '#f28b82',
-    yellow: '#fdd663',
-    green: '#81c995',
-    pink: '#ff8bcb',
-    purple: '#c58af9',
-    cyan: '#78d9ec',
-    orange: '#ffab70'
-};
-
-function hexToRgba(hex, alpha) {
-    const r = parseInt(hex.slice(1, 3), 16);
-    const g = parseInt(hex.slice(3, 5), 16);
-    const b = parseInt(hex.slice(5, 7), 16);
-    return `rgba(${r}, ${g}, ${b}, ${alpha})`;
-}
 
 export function createTabElement(tab, { onAddToGroupClick }) {
     const tabItem = document.createElement('div');
